refactor: components/common 디렉토리명을 ui 로 변경 (#1)
